+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix `sourceid2nslc()` to be honest about return type and accurate docs.
 - Fix return type for `MS3Record.encoding_str()`, will never be None.
 - Add ownership test to `MS3Record.parse_into()` to avoid clobbering foreign data.
+- Harden `MS3Record.packet()` and `MS3Record.generate()` by raising an exception
+  when only one of `data_samples` or `sample_type` is provided.
 - Type annotations converted to PEP 604 / PEP 585 form.
 - Apply consistent python formatting with ruff.